test(batch): cover compressed batches on the fetch adapter
Replaces the socketless variant, so the client is an ordinary RPCLink whose
fetch decompresses the response on its own.
